functional radio signal attenuation

+++ src/ATC/trafficcontrol.hxx
@@ -304,7 +304,7 @@ public:
   string getGateName(FGAIAircraft *aircraft);
   virtual void render(bool) = 0;
   virtual string getName()  = 0;
-  int calculate_attenuation(FGTrafficRecord * rec, FGAirportDynamics *parent, int ground_to_air);
+  double calculate_attenuation(FGTrafficRecord * rec, FGAirportDynamics *parent, int ground_to_air);
 
 private:
